apps/api/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic-settings
azure-search-documents
google-generativeai
pypdf
python-multipart
httpx
